feat: added composition + convert float to double for price

+++ src/jp_lab5_gui/Classes/Bill.java
@@ -3,7 +3,7 @@ package jp_lab5_gui;
 
 public class Bill {
   public String acc_holder;
-  public float amount;
+  public double amount;
 
   public String getAccHolders() { 
     return acc_holder;
